solution: Fixed project 'Battle ground'
